# specs/001-job-matching-system/contracts/detailed-api-spec.yaml
openapi: 3.0.3
info:
  title: Job Matching System - Detailed API Specification
  description: |
    バイト求人マッチングシステムの詳細API仕様書

    ## システム概要
    - 10万件の求人データから1万人のユーザーに最適な40件を毎日選定
    - パーソナライズされたメール配信準備
    - 3段階スコアリング（基礎・SEO・パーソナライズ）
    - 6セクション構成のメール生成

    ## 技術スタック
    - Backend: Python 3.11 + FastAPI
    - Database: Supabase (PostgreSQL)
    - AI: OpenAI GPT-4/5 nano
    - Monitoring: Real-time SQL dashboard
  version: 2.0.0
  contact:
    name: Job Matching System Team
    email: dev@job-matching.com
    url: https://github.com/job-matching/api
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v2
    description: Development server
  - url: https://staging-api.job-matching.com/v2
    description: Staging server
  - url: https://api.job-matching.com/v2
    description: Production server
...
paths:
...
  /jobs:
    get:
      tags: [Jobs]
      summary: 求人一覧取得
      description: |
        求人データの一覧を取得。
        検索、フィルタリング、ページネーション、ソート機能付き。
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: ページ番号
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 50
          description: 1ページあたりの件数
        - name: keyword
          in: query
          schema:
            type: string
            maxLength: 100
          description: キーワード検索（職種、会社名、勤務地など）
        - name: location
          in: query
          schema:
            type: string
          description: 勤務地フィルタ（部分一致）
        - name: salary_min
          in: query
          schema:
            type: integer
            minimum: 0
          description: 最低時給（円）
        - name: salary_max
          in: query
          schema:
            type: integer
            minimum: 0
          description: 最高時給（円）
        - name: occupation_categories
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [restaurant, retail, office, logistics, education, medical,
                     construction, manufacturing, it, service, creative, other]
          style: form
          explode: true
          description: 職種カテゴリフィルタ
        - name: needs_categories
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [daily_payment, high_hourly, short_term, long_term, student_ok,
                     unexperienced_ok, transportation_provided, meal_provided,
                     uniform_provided, near_station, flexible_shift, weekend_only,
                     night_shift, remote_ok]
          style: form
          explode: true
          description: ニーズカテゴリフィルタ
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, paused, expired, deleted]
            default: active
          description: 求人ステータス
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, salary, basic_score, seo_score,
                   personalized_score, combined_score]
            default: updated_at
          description: ソート項目
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: ソート順
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
          description: 作成日時フィルタ（以降）
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
          description: 作成日時フィルタ（以前）
      responses:
        '200':
          description: 求人一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Jobs]
      summary: 求人作成
      description: 新しい求人を作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '201':
          description: 求人作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: 求人詳細取得
      description: 指定されたIDの求人詳細を取得
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
          description: 求人ID
      responses:
        '200':
          description: 求人詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Jobs]
      summary: 求人更新
      description: 指定されたIDの求人を更新
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
          description: 求人ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobUpdateRequest'
      responses:
        '200':
          description: 求人更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Jobs]
      summary: 求人削除
      description: 指定されたIDの求人を削除（論理削除）
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: integer
          description: 求人ID
      responses:
        '204':
          description: 求人削除成功
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /jobs/search:
    post:
      tags: [Jobs]
      summary: 高度な求人検索
      description: |
        複雑な検索条件による求人検索
        全文検索、地理的検索、スコア範囲指定など
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobSearchRequest'
      responses:
        '200':
          description: 検索結果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /batch/jobs:
    get:
      tags: [Batch]
      summary: バッチジョブ一覧取得
      description: |
        実行中または完了したバッチジョブの一覧を取得
        ステータス監視とログ確認用
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: array
            items:
              type: string
              enum: [pending, running, completed, failed, cancelled]
          style: form
          explode: true
          description: ステータスフィルタ
        - name: job_type
          in: query
          schema:
            type: string
            enum: [data_import, scoring, categorization, matching, email_generation, all]
          description: ジョブ種別フィルタ
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, started_at, completed_at]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: バッチジョブ一覧
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Batch]
      summary: バッチジョブ作成
      description: |
        新しいバッチジョブを作成・スケジュール
        各種処理の一括実行管理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchJobCreateRequest'
      responses:
        '201':
          description: バッチジョブ作成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /batch/jobs/{job_id}:
    get:
      tags: [Batch]
      summary: バッチジョブ詳細取得
      description: 指定されたバッチジョブの詳細とログを取得
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: バッチジョブID
        - name: include_logs
          in: query
          schema:
            type: boolean
            default: true
          description: ログを含めるか
        - name: log_level
          in: query
          schema:
            type: string
            enum: [debug, info, warning, error]
            default: info
          description: ログレベルフィルタ
      responses:
        '200':
          description: バッチジョブ詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJobDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags: [Batch]
      summary: バッチジョブ更新
      description: バッチジョブの設定を更新（実行前のみ）
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchJobUpdateRequest'
      responses:
        '200':
          description: バッチジョブ更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchJob'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 実行中のジョブは更新不可
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: [Batch]
      summary: バッチジョブキャンセル
      description: 実行中のバッチジョブをキャンセル
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: キャンセル成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:
                    type: string
                  status:
                    type: string
                    enum: [cancelled]
                  cancelled_at:
                    type: string
                    format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: キャンセル不可能な状態
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total_items:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
        has_next:
          type: boolean
        has_prev:
          type: boolean
...
    Job:
      type: object
      properties:
        id:
          type: integer
        external_id:
          type: string
          description: 外部システムの求人ID
        company_name:
          type: string
        title:
          type: string
        description:
          type: string
        salary_min:
          type: integer
          description: 最低時給（円）
        salary_max:
          type: integer
          description: 最高時給（円）
        location:
          type: string
        prefecture:
          type: string
        city:
          type: string
        nearest_station:
          type: string
        work_hours:
          type: string
        work_days:
          type: string
        features:
          type: array
          items:
            type: string
          description: 求人特徴（日払い、高時給など）
        occupation_category:
          type: string
          enum: [restaurant, retail, office, logistics, education, medical,
                 construction, manufacturing, it, service, creative, other]
        needs_categories:
          type: array
          items:
            type: string
            enum: [daily_payment, high_hourly, short_term, long_term, student_ok,
                   unexperienced_ok, transportation_provided, meal_provided,
                   uniform_provided, near_station, flexible_shift, weekend_only,
                   night_shift, remote_ok]
        basic_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        seo_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        personalized_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        combined_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        status:
          type: string
          enum: [draft, active, paused, expired, deleted]
        url:
          type: string
          format: uri
        application_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
      required:
        - id
        - company_name
        - title
        - status
    JobCreateRequest:
      type: object
      properties:
        external_id:
          type: string
        company_name:
          type: string
          maxLength: 100
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        salary_min:
          type: integer
          minimum: 0
        salary_max:
          type: integer
          minimum: 0
        location:
          type: string
          maxLength: 200
        prefecture:
          type: string
        city:
          type: string
        nearest_station:
          type: string
        work_hours:
          type: string
        work_days:
          type: string
        features:
          type: array
          items:
            type: string
        occupation_category:
          type: string
          enum: [restaurant, retail, office, logistics, education, medical,
                 construction, manufacturing, it, service, creative, other]
        url:
          type: string
          format: uri
        application_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
      required:
        - company_name
        - title
        - occupation_category
    JobUpdateRequest:
      type: object
      properties:
        company_name:
          type: string
          maxLength: 100
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        salary_min:
          type: integer
          minimum: 0
        salary_max:
          type: integer
          minimum: 0
        location:
          type: string
          maxLength: 200
        work_hours:
          type: string
        work_days:
          type: string
        features:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, active, paused, expired]
        url:
          type: string
          format: uri
        application_url:
          type: string
          format: uri
        expires_at:
          type: string
          format: date-time
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
        filters:
          type: object
          description: 適用されたフィルタ情報
    JobSearchRequest:
      type: object
      properties:
        query:
          type: string
          maxLength: 200
          description: 全文検索クエリ
        location:
          type: object
          properties:
            point:
              type: object
              properties:
                lat:
                  type: number
                  format: float
                lng:
                  type: number
                  format: float
            radius_km:
              type: number
              format: float
              minimum: 0
              maximum: 100
          description: 地理的検索
        salary_range:
          type: object
          properties:
            min:
              type: integer
              minimum: 0
            max:
              type: integer
              minimum: 0
        score_range:
          type: object
          properties:
            score_type:
              type: string
              enum: [basic, seo, personalized, combined]
            min:
              type: number
              format: float
              minimum: 0
              maximum: 100
            max:
              type: number
              format: float
              minimum: 0
              maximum: 100
        filters:
          type: object
          properties:
            occupation_categories:
              type: array
              items:
                type: string
            needs_categories:
              type: array
              items:
                type: string
            status:
              type: array
              items:
                type: string
        sort:
          type: object
          properties:
            field:
              type: string
              enum: [created_at, updated_at, salary, basic_score, seo_score,
                     personalized_score, combined_score, relevance]
            order:
              type: string
              enum: [asc, desc]
        page:
          type: integer
          minimum: 1
          default: 1
        limit:
          type: integer
          minimum: 1
          maximum: 1000
          default: 50
    JobSearchResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Job'
              - type: object
                properties:
                  relevance_score:
                    type: number
                    format: float
                  highlight:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: string
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
        search_meta:
          type: object
          properties:
            total_time_ms:
              type: integer
            query_analyzed:
              type: string
            suggestions:
              type: array
              items:
                type: string
...
  responses:
    BadRequest:
      description: リクエストが無効です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: アクセス権限がありません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        APIキー認証
        ヘッダーに `X-API-Key: your-api-key` を含めてください

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT認証
        ヘッダーに `Authorization: Bearer <token>` を含めてください
security:
  - ApiKeyAuth: []
  - BearerAuth: []

# ===========================================
# 追加情報・拡張仕様
# ===========================================
...